WeatherApp/requirements.txt
nose
pyweatherbit
pathlib
validate_email
